many fixes

+++ src/test/java/io/supertokens/storage/postgresql/test/TestingProcessManager.java
@@ -129,14 +129,14 @@ public class TestingProcessManager {
             return args;
         }
 
-        public void kill() throws InterruptedException {
+        public void kill(boolean removeAllInfo) throws InterruptedException {
             if (killed) {
                 return;
             }
             // we check if there are multiple user pool IDs loaded, and if there are,
             // we clear all the info before killing cause otherwise those extra dbs will retain info
             // across tests
-            if (StorageLayer.hasMultipleUserPools(this.main)) {
+            if (removeAllInfo && StorageLayer.hasMultipleUserPools(this.main)) {
                 try {
                     main.deleteAllInformationForTesting();
                 } catch (Exception e) {
@@ -149,6 +149,10 @@ public class TestingProcessManager {
             }
             main.killForTestingAndWaitForShutdown();
             killed = true;
+        }
+
+        public void kill() throws InterruptedException {
+            kill(true);
         }
 
         public EventAndException checkOrWaitForEvent(PROCESS_STATE state) throws InterruptedException {
